Update and refactoring of classes
Updated names in order for merge to be compatible.
