+++ .idea/workspace.xml
@@ -6,22 +6,8 @@
   <component name="ChangeListManager">
     <list default="true" id="d7470bda-e2b2-4312-9e1d-70282ac9a8ab" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/Black_Bishop.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/Black_Bishop.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/Black_King.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/Black_King.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/Black_Knight.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/Black_Knight.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/Black_Pawn.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/Black_Pawn.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/Black_Queen.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/Black_Queen.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/Black_Rook.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/Black_Rook.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/White_Bishop.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/White_Bishop.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/White_King.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/White_King.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/White_Knight.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/White_Knight.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/White_Pawn.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/White_Pawn.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/White_Queen.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/White_Queen.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Assets/White_Rook.png" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/PieceImages/White_Rook.png" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/MVC/Menu/mainMenu.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/MVC/Menu/MainMenu.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/MVC/controller/MenuController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/MVC/controller/MenuController.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/resources/src/main/resources/chessBackground.jpeg" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/MenuImages/chessBackground.jpeg" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/resources/test2.fxml" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/MainMenu.fxml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/MVC/Menu/MainMenu.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/MVC/view/MainMenu.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/module-info.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/module-info.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -70,11 +56,11 @@
   </component>
   <component name="RunManager" selected="Application.MainMenu">
     <configuration name="MainMenu" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="MVC.Menu.MainMenu" />
+      <option name="MAIN_CLASS_NAME" value="MVC.view.MainMenu" />
       <module name="CrazyChess-DIT213" />
       <extension name="coverage">
         <pattern>
-          <option name="PATTERN" value="MVC.Menu.*" />
+          <option name="PATTERN" value="MVC.view.*" />
           <option name="ENABLED" value="true" />
         </pattern>
       </extension>
