+++ src/main/java/module-info.java
@@ -2,8 +2,6 @@ module MVC {
     requires javafx.graphics;
     requires javafx.controls;
     requires javafx.fxml;
-    exports MVC.Menu;
     exports MVC.controller;
-
-
+    exports MVC.view;
 }
